Implement discussed improvements - New method CountDistinct
